Lots of general improvements. Changed scalar values to be doubles. Added a little more pre-processing regexes to the main class and added a special exception for simple 24hr format times.
git-svn-id: http://goda-time.googlecode.com/svn/trunk@12 6e3c4dfa-b0e6-11dd-a5e3-f1c34dd4e8bc

+++ src/org/goda/chronic/repeaters/Repeater.java
@@ -84,7 +84,7 @@ public abstract class Repeater<T> extends Tag<T> implements Comparable<Repeater<
 
   protected abstract MutableInterval _thisMutableInterval(Pointer.PointerType pointer);
 
-  public abstract MutableInterval getOffset(MutableInterval MutableInterval, int amount, Pointer.PointerType pointer);
+  public abstract MutableInterval getOffset(MutableInterval MutableInterval, double amount, Pointer.PointerType pointer);
 
   @Override
   public String toString() {
