+++ src/org/goda/chronic/repeaters/RepeaterHour.java
@@ -56,10 +56,10 @@ public class RepeaterHour extends RepeaterUnit {
   }
 
   @Override
-  public MutableInterval getOffset(MutableInterval span, int amount, Pointer.PointerType pointer) {
+  public MutableInterval getOffset(MutableInterval span, double amount, Pointer.PointerType pointer) {
     int direction = (pointer == Pointer.PointerType.FUTURE) ? 1 : -1;
     // WARN: Does not use Calendar
-    int seconds = direction * amount * RepeaterHour.HOUR_SECONDS;
+    int seconds = (int) (direction * amount * RepeaterHour.HOUR_SECONDS);
     return new MutableInterval( span.getStart().plusSeconds(seconds), span.getEnd().plusSeconds(seconds));
     
   }
