+++ src/org/goda/chronic/repeaters/RepeaterWeek.java
@@ -79,10 +79,10 @@ public class RepeaterWeek extends RepeaterUnit {
   }
 
   @Override
-  public MutableInterval getOffset(MutableInterval span, int amount, Pointer.PointerType pointer) {
+  public MutableInterval getOffset(MutableInterval span, double amount, Pointer.PointerType pointer) {
     int direction = (pointer == Pointer.PointerType.FUTURE) ? 1 : -1;
     // WARN: Does not use Calendar
-    int seconds = direction * amount * RepeaterWeek.WEEK_SECONDS;
+    int seconds = (int) (direction * amount * RepeaterWeek.WEEK_SECONDS);
     return new MutableInterval( span.getStart().plusSeconds(seconds), span.getEnd().plusSeconds(seconds));
   }
